requirements.txt
mysql-connector-python==9.1.0
argon2-cffi==23.1.0      
argon2-cffi-bindings==21.2.0  
sqlparse==0.5.0
protobuf==3.20.1
pycryptodome==3.14.1
Pygments==2.12.0
commonmark==0.9.1
pyperclip==1.8.2
rich==12.3.0
